add: useKakaoInit
