+++ packages/santa_close_frame/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useKakaoInit'
